// src/service/engine/v1/types.ts
export type ID = { engine_id: string; account_id: string };

export type Engine = {
  id: ID;
  name: string;
  description: string;
  endpoint: string;
  current_status_summary: EngineStatusSummary;
};

export enum EngineStatusSummary {
  DELETING = "ENGINE_STATUS_SUMMARY_DELETING",
  REPAIRING = "ENGINE_STATUS_SUMMARY_REPAIRING",
  RUNNING = "ENGINE_STATUS_SUMMARY_RUNNING",
  STARTING = "ENGINE_STATUS_SUMMARY_STARTING",
  STARTING_INITIALIZING = "ENGINE_STATUS_SUMMARY_STARTING_INITIALIZING",
  STOPPED = "ENGINE_STATUS_SUMMARY_STOPPED",
  STOPPING = "ENGINE_STATUS_SUMMARY_STOPPING",
  // Specific for V1
  DELETED = "ENGINE_STATUS_SUMMARY_DELETED",
  FAILED = "ENGINE_STATUS_SUMMARY_FAILED",
  RESTARTING = "ENGINE_STATUS_SUMMARY_RESTARTING",
  RESTARTING_INITIALIZING = "ENGINE_STATUS_SUMMARY_RESTARTING_INITIALIZING",
  UNSPECIFIED = "ENGINE_STATUS_SUMMARY_UNSPECIFIED",
  UPGRADING = "ENGINE_STATUS_SUMMARY_UPGRADING"
}